int -> NSInteger

+++ REUIKitAdditions/UIImage+REUIKitAdditions.h
@@ -28,6 +28,6 @@
 @interface UIImage (REUIKitAdditions)
 
 + (UIImage *)maskedImageNamed:(NSString *)name color:(UIColor *)color;
-+ (UIImage *)squareImageWithImage:(UIImage *)image size:(int)size;
++ (UIImage *)squareImageWithImage:(UIImage *)image size:(NSInteger)size;
 
 @end
